refactor: move test_io fixtures to conftest.py
