+++ tests/conftest.py
@@ -1,7 +1,28 @@
 from enum import Enum
 
+import numpy as np
+import PIL.Image as PIL_Image
 import pytest
 from pydantic import BaseModel
+
+
+@pytest.fixture
+def sample_nd_array_data():
+    return np.zeros((100, 100, 3), dtype=np.uint8)
+
+
+@pytest.fixture
+def sample_audio_data():
+    rate = 16000
+    return np.random.uniform(-1, 1, rate)
+
+
+@pytest.fixture
+def sample_pil_image_data() -> PIL_Image.Image:
+    width, height = 100, 100
+    nd_image = np.zeros((height, width, 3), dtype=np.uint8)
+    pil_image = PIL_Image.fromarray(nd_image)
+    return pil_image
 
 
 # test_utils fixtures.
