+++ tests/test_io.py
@@ -8,25 +8,6 @@ from hyko_sdk.io import CSV, PDF, Audio, Image, Video
 from hyko_sdk.models import Ext
 
 
-@pytest.fixture
-def sample_nd_array_data():
-    return np.zeros((100, 100, 3), dtype=np.uint8)
-
-
-@pytest.fixture
-def sample_audio_data():
-    rate = 16000
-    return np.random.uniform(-1, 1, rate)
-
-
-@pytest.fixture
-def sample_pil_image_data() -> PIL_Image.Image:
-    width, height = 100, 100
-    nd_image = np.zeros((height, width, 3), dtype=np.uint8)
-    pil_image = PIL_Image.fromarray(nd_image)
-    return pil_image
-
-
 ############## Image Tests #####################
 @pytest.mark.parametrize(
     "name",
